tidy and clean up old tracking code
